Eslint fixes and proptype implemented
