+++ src/containers/Attendance/mock-data/fiveDayData.json
@@ -1,0 +1,53 @@
+{
+    "pastFiveDays": [
+        {
+          "date": "2018-08-06",
+          "timesStamp": {
+            "time_in": "08:46:25",
+            "time_out": "17:46:25",
+            "late": false,
+            "left_early": false
+          },
+          "attendance": "full",
+          "index": 1
+        },
+        {
+          "date": "2018-08-07",
+          "timesStamp": {
+            "time_in": "08:46:25",
+            "time_out": "17:46:25",
+            "late": false,
+            "left_early": false
+          },
+          "attendance": "full",
+          "index": 1
+        },
+        {
+          "date": "2018-08-08",
+          "timesStamp": {
+            "time_in": "10:46:25",
+            "time_out": "14:46:25",
+            "late": true,
+            "left_early": true
+          },
+          "attendance": "partial",
+          "index": 1
+        },
+        {
+          "date": "2018-08-09",
+          "timesStamp": {
+            "time_in": "11:16:25",
+            "time_out": "17:02:25",
+            "late": true,
+            "left_early": false
+          },
+          "attendance": "partial",
+          "index": 1
+        },
+        {
+          "date": "2018-08-10",
+          "attendance": "absent",
+          "index": 1
+        }
+      ]
+}
